refactor(core): extract utils into seperate package
