+++ alvoc/cli.py
@@ -1,7 +1,7 @@
 import typer
 
 from alvoc.core.logging import init_logger
-from alvoc.core.mutations import aa, nt
+from alvoc.core.utils.convert import aa, nt
 from alvoc.core.mutations.analyze import find_mutants as find_mutants_internal
 from alvoc.core.precompute import precompute
 # from alvoc.core.lineages import find_lineages
